Uno.Data.Xml: reorganize files
