--- lib/Uno.Data.Xml/Tests/Xml.Test.unosln
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<Solution>
-  <Project Path="Xml.Test/Xml.Test.unoproj" />
-</Solution>
