+++ lib/Uno.Data.Xml/Uno.Data.Xml.unoproj
@@ -6,35 +6,35 @@
     "../Uno.Collections/Uno.Collections.unoproj"
   ],
   "Includes": [
-    "Targets/CIL/Uno.Data.Xml.cil.uxl:Extensions",
-    "Targets/CPlusPlus/Uno.Data.Xml.cpp.uxl:Extensions",
-    "Targets/CPlusPlus/Source/pugiconfig.hpp:File",
-    "Targets/CPlusPlus/Source/pugixml.cpp:File",
-    "Targets/CPlusPlus/Source/pugixml.hpp:File",
-    "Source/System.Xml.uno:Source",
-    "Source/XmlException.uno:Source",
-    "Source/XmlNodeImpl.uno:Source",
-    "Source/XmlNodeHandle.uno:Source",
-    "Source/XmlAttributeHandle.uno:Source",
-    "Source/XmlEnums.uno:Source",
-    "Source/Elements/XmlAttribute.uno:Source",
-    "Source/Elements/XmlValue.uno:Source",
-    "Source/Elements/XmlText.uno:Source",
-    "Source/Elements/XmlProcessingInstruction.uno:Source",
-    "Source/Elements/XmlNode.uno:Source",
-    "Source/Elements/XmlLinkedNode.uno:Source",
-    "Source/Elements/XmlElement.uno:Source",
-    "Source/Elements/XmlDocumentElement.uno:Source",
-    "Source/Elements/XmlDocument.uno:Source",
-    "Source/Elements/XmlDeclaration.uno:Source",
-    "Source/Elements/XmlComment.uno:Source",
-    "Source/Elements/XmlCharacterData.uno:Source",
-    "Source/Collections/XmlNodeCollectionBase.uno:Source",
-    "Source/Collections/XmlAttributeCollection.uno:Source",
-    "Source/Collections/XmlLinkedNodeCollection.uno:Source",
-    "Source/Helpers/AttributeHandleConverter.uno:Source",
-    "Source/Helpers/ValidationHelper.uno:Source",
-    "Source/Helpers/XmlEncodingHelper.uno:Source",
-    "Source/Helpers/XmlNodeHandleConverter.uno:Source"
+    "cil/System.Xml.uno:Source",
+    "cil/System.Xml.uxl:Extensions",
+    "cpp/pugixml.uxl:Extensions",
+    "cpp/pugiconfig.hpp:File",
+    "cpp/pugixml.cpp:File",
+    "cpp/pugixml.hpp:File",
+    "XmlException.uno:Source",
+    "XmlNodeImpl.uno:Source",
+    "XmlNodeHandle.uno:Source",
+    "XmlAttributeHandle.uno:Source",
+    "XmlEnums.uno:Source",
+    "elements/XmlAttribute.uno:Source",
+    "elements/XmlValue.uno:Source",
+    "elements/XmlText.uno:Source",
+    "elements/XmlProcessingInstruction.uno:Source",
+    "elements/XmlNode.uno:Source",
+    "elements/XmlLinkedNode.uno:Source",
+    "elements/XmlElement.uno:Source",
+    "elements/XmlDocumentElement.uno:Source",
+    "elements/XmlDocument.uno:Source",
+    "elements/XmlDeclaration.uno:Source",
+    "elements/XmlComment.uno:Source",
+    "elements/XmlCharacterData.uno:Source",
+    "collections/XmlNodeCollectionBase.uno:Source",
+    "collections/XmlAttributeCollection.uno:Source",
+    "collections/XmlLinkedNodeCollection.uno:Source",
+    "helpers/AttributeHandleConverter.uno:Source",
+    "helpers/ValidationHelper.uno:Source",
+    "helpers/XmlEncodingHelper.uno:Source",
+    "helpers/XmlNodeHandleConverter.uno:Source"
   ]
 }
