+++ lib/Uno.Data.Xml/collections/XmlLinkedNodeCollection.uno
