+++ lib/Uno.Data.Xml/cpp/pugixml.cpp
@@ -14,7 +14,7 @@
 #ifndef SOURCE_PUGIXML_CPP
 #define SOURCE_PUGIXML_CPP
 
-#include "Source/pugixml.hpp"
+#include "pugixml.hpp"
 
 #include <stdlib.h>
 #include <stdio.h>
