+++ lib/Uno.Data.Xml/elements/XmlDeclaration.uno
