+++ lib/Uno.Data.Xml/elements/XmlValue.uno
