+++ lib/Uno.Data.Xml/helpers/XmlEncodingHelper.uno
