+++ lib/Uno.Data.Xml/helpers/XmlNodeHandleConverter.uno
